@@ -18,7 +18,8 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 0.0.1+1
 
 environment:
-  sdk: ">=3.0.6"
+  sdk: ">=3.1.0"
+  flutter: ^3.10.0
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -52,7 +53,9 @@ dependencies:
   flutter_keyboard_visibility: ^5.4.0
   expandable_text: ^2.3.0
   cached_network_image: ^3.2.1
-  flutter_mentions: ^2.0.1
+  flutter_mentions:
+    git:
+      url: https://github.com/gtalha07/flutter_mentions.git
   flutter_matrix_html:
     git:
       url: https://github.com/acterglobal/flutter_matrix_html.git
@@ -128,7 +131,7 @@ dev_dependencies:
   icons_launcher: ^2.0.7
   flutter_lints: ^2.0.2
   build_runner: ^2.3.3
-  mockito: ^5.0.0
+  mockito: ^5.4.2
   # The current build runner version is breaking when generating due to the issue 
   # https://github.com/dart-lang/build/issues/3519, the most probable fix is to upgrade
   # but that requires updated flutter SDK, temporarily we can add build_runner_core
@@ -176,7 +179,6 @@ flutter:
   assets:
     - assets/images/
     - assets/icon/
-    - google_fonts/
 
   #   - images/a_dot_ham.jpeg
 
@@ -191,7 +193,13 @@ flutter:
   # "family" key with the font family name, and a "fonts" key with a
   # list giving the asset and other descriptors for the font. For
   # example:
-  # fonts:
+  fonts:
+    - family: Inter
+      fonts:
+        - asset: assets/fonts/Inter.ttf
+    - family: NotoEmoji
+      fonts:
+        - asset: assets/fonts/Noto-COLRv1.ttf
   #   - family: Schyler
   #     fonts:
   #       - asset: fonts/Schyler-Regular.ttf
