@@ -1,25 +1,42 @@
-import 'dart:io';
+import 'dart:math';
 
+import 'package:acter/common/themes/app_theme.dart';
 import 'package:emoji_picker_flutter/emoji_picker_flutter.dart';
-import 'package:expandable_text/expandable_text.dart';
-import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter/material.dart';
 
 class EmojiPickerWidget extends StatelessWidget {
-  const EmojiPickerWidget({Key? key, required this.onSelected})
-      : super(key: key);
+  const EmojiPickerWidget({
+    Key? key,
+    this.size,
+    this.onEmojiSelected,
+    this.onBackspacePressed,
+    this.withBoarder = false,
+  }) : super(key: key);
 
-  final StringCallback onSelected;
+  final Size? size;
+  final bool withBoarder;
+  final OnEmojiSelected? onEmojiSelected;
+  final OnBackspacePressed? onBackspacePressed;
 
   @override
   Widget build(BuildContext context) {
+    final height =
+        size == null ? MediaQuery.of(context).size.height / 3 : size!.height;
+    final width =
+        size == null ? MediaQuery.of(context).size.width : size!.width;
+    final cols = min(width / (EmojiConfig.emojiSizeMax * 2), 12).floor();
+
     return Container(
-      padding: const EdgeInsets.only(top: 10, left: 15, right: 15),
-      decoration: const BoxDecoration(
-        color: Colors.white,
-        borderRadius: BorderRadius.vertical(top: Radius.circular(16)),
-      ),
-      height: MediaQuery.of(context).size.height / 2,
+      padding: withBoarder
+          ? const EdgeInsets.only(top: 10, left: 15, right: 15)
+          : null,
+      decoration: withBoarder
+          ? const BoxDecoration(
+              color: Colors.white,
+              borderRadius: BorderRadius.vertical(top: Radius.circular(16)),
+            )
+          : null,
+      height: height,
       child: Column(
         children: [
           Container(
@@ -33,14 +50,16 @@ class EmojiPickerWidget extends StatelessWidget {
           ),
           Expanded(
             child: EmojiPicker(
-              onEmojiSelected: (Category? category, Emoji emoji) =>
-                  onSelected(emoji.emoji),
+              onEmojiSelected: onEmojiSelected,
+              onBackspacePressed: onBackspacePressed,
               config: Config(
-                columns: 7,
-                emojiSizeMax: 32 * ((!kIsWeb && Platform.isIOS) ? 1.30 : 1.0),
+                columns: cols,
+                checkPlatformCompatibility:
+                    EmojiConfig.checkPlatformCompatibility,
+                emojiSizeMax: EmojiConfig.emojiSizeMax,
                 initCategory: Category.RECENT,
-                bgColor: Colors.white,
-                recentTabBehavior: RecentTabBehavior.NONE,
+                bgColor: Theme.of(context).colorScheme.background,
+                recentTabBehavior: RecentTabBehavior.RECENT,
                 recentsLimit: 28,
               ),
             ),
