@@ -1,3 +1,4 @@
+import 'package:acter/common/themes/app_theme.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_ui/flutter_chat_ui.dart';
 
@@ -49,7 +50,11 @@ class ActerChatTheme extends ChatTheme {
     double messageInsetsHorizontal = 20,
     double messageInsetsVertical = 16,
     Color primaryColor = const Color(0xffFF8E00),
-    TextStyle receivedEmojiMessageTextStyle = const TextStyle(fontSize: 40),
+    TextStyle receivedEmojiMessageTextStyle = const TextStyle(
+      fontSize: 40,
+      fontFamily: emojiFont,
+      fontFamilyFallback: emojiFallbackFonts,
+    ),
     TextStyle? receivedMessageBodyBoldTextStyle,
     TextStyle? receivedMessageBodyCodeTextStyle,
     TextStyle? receivedMessageBodyLinkTextStyle,
@@ -83,7 +88,11 @@ class ActerChatTheme extends ChatTheme {
     Widget? sendButtonIcon,
     EdgeInsets? sendButtonMargin,
     Widget? sendingIcon,
-    TextStyle sentEmojiMessageTextStyle = const TextStyle(fontSize: 40),
+    TextStyle sentEmojiMessageTextStyle = const TextStyle(
+      fontSize: 40,
+      fontFamily: emojiFont,
+      fontFamilyFallback: emojiFallbackFonts,
+    ),
     TextStyle? sentMessageBodyBoldTextStyle,
     TextStyle? sentMessageBodyCodeTextStyle,
     TextStyle? sentMessageBodyLinkTextStyle,
