@@ -191,7 +191,7 @@ class _CreateEventSheetConsumerState extends ConsumerState<CreateEventSheet> {
                         ),
                       ],
                     ),
-                  )
+                  ),
                 ],
               ),
               const SizedBox(height: 15),
@@ -325,6 +325,9 @@ class _CreateEventSheetConsumerState extends ConsumerState<CreateEventSheet> {
         pathParameters: {'calendarId': eventId.toString()},
       );
     } catch (e) {
+      if (!context.mounted) {
+        return;
+      }
       context.pop();
       customMsgSnackbar(context, 'Some error occurred $e');
     }
