@@ -4,6 +4,7 @@ import 'package:acter/common/themes/app_theme.dart';
 import 'package:acter/common/utils/constants.dart';
 import 'package:acter/common/widgets/user_avatar.dart';
 import 'package:acter/features/home/providers/client_providers.dart';
+import 'package:acter/features/home/widgets/in_dashboard.dart';
 import 'package:acter/features/home/widgets/my_spaces_section.dart';
 import 'package:acter/features/home/widgets/my_events.dart';
 import 'package:acter/features/settings/providers/settings_providers.dart';
@@ -106,64 +107,55 @@ class _DashboardState extends ConsumerState<Dashboard> {
         ),
       ];
     }
-    return Scaffold(
-      body: Container(
-        decoration: BoxDecoration(
-          gradient: LinearGradient(
-            begin: Alignment.topCenter,
-            end: Alignment.center,
-            colors: <Color>[
-              Theme.of(context).colorScheme.background,
-              Theme.of(context).colorScheme.neutral,
+    return InDashboard(
+      child: Scaffold(
+        body: Container(
+          decoration: const BoxDecoration(gradient: AppTheme.primaryGradient),
+          child: CustomScrollView(
+            slivers: <Widget>[
+              SliverAppBar(
+                backgroundColor: Colors.transparent,
+                actions: <Widget>[
+                  IconButton(
+                    icon: const Icon(Atlas.settings_monitor_thin),
+                    onPressed: () {
+                      customMsgSnackbar(
+                        context,
+                        'Configuration Page for Dashboard not yet implemented',
+                      );
+                    },
+                  ),
+                  IconButton(
+                    icon: const Icon(Atlas.construction_tools_thin),
+                    onPressed: () => context.pushNamed(Routes.settings.name),
+                  ),
+                  Visibility(
+                    // FIXME: Only show mobile / when bottom bar shown...
+                    visible: !client.isGuest(),
+                    replacement: InkWell(
+                      onTap: () => context.pushNamed(Routes.authLogin.name),
+                      child: ActerAvatar(
+                        uniqueId: UniqueKey().toString(),
+                        mode: DisplayMode.User,
+                      ),
+                    ),
+                    child: Container(
+                      key: Keys.avatar,
+                      margin: const EdgeInsets.all(8),
+                      child: InkWell(
+                        onTap: () => context.pushNamed(Routes.myProfile.name),
+                        child: const UserAvatarWidget(),
+                      ),
+                    ),
+                  ),
+                ],
+                title: isDesktop
+                    ? const Text('My Dashboard')
+                    : const Text('Overview'),
+              ),
+              ...children,
             ],
           ),
-        ),
-        child: CustomScrollView(
-          slivers: <Widget>[
-            SliverAppBar(
-              backgroundColor: Colors.transparent,
-              actions: <Widget>[
-                IconButton(
-                  icon: const Icon(Atlas.settings_monitor_thin),
-                  onPressed: () {
-                    customMsgSnackbar(
-                      context,
-                      'Configuration Page for Dashboard not yet implemented',
-                    );
-                  },
-                ),
-                IconButton(
-                  icon: const Icon(Atlas.construction_tools_thin),
-                  onPressed: () {
-                    context.go('/settings');
-                  },
-                ),
-                Visibility(
-                  // FIXME: Only show mobile / when bottom bar shown...
-                  visible: !client.isGuest(),
-                  replacement: InkWell(
-                    onTap: () => context.pushNamed(Routes.authLogin.name),
-                    child: ActerAvatar(
-                      uniqueId: UniqueKey().toString(),
-                      mode: DisplayMode.User,
-                    ),
-                  ),
-                  child: Container(
-                    key: Keys.avatar,
-                    margin: const EdgeInsets.all(8),
-                    child: InkWell(
-                      onTap: () => context.pushNamed(Routes.myProfile.name),
-                      child: const UserAvatarWidget(),
-                    ),
-                  ),
-                ),
-              ],
-              title: isDesktop
-                  ? const Text('My Dashboard')
-                  : const Text('Overview'),
-            ),
-            ...children,
-          ],
         ),
       ),
     );
