@@ -70,7 +70,7 @@ class SpaceOverview extends ConsumerWidget {
                 RelatedSpacesCard(spaceId: spaceIdOrAlias),
               ],
             ),
-          )
+          ),
         ],
       ),
     );
