@@ -1,8 +1,8 @@
 import 'package:acter/common/dialogs/dialog_page.dart';
 import 'package:acter/common/dialogs/side_sheet_page.dart';
+import 'package:acter/common/themes/app_theme.dart';
 import 'package:acter/common/utils/constants.dart';
 import 'package:acter/common/utils/routes.dart';
-import 'package:acter/common/utils/utils.dart';
 import 'package:acter/features/activities/pages/activities_page.dart';
 import 'package:acter/features/activities/pages/sessions_page.dart';
 import 'package:acter/features/bug_report/pages/bug_report_page.dart';
@@ -345,28 +345,6 @@ List<RouteBase> makeRoutes(Ref ref) {
 
     GoRoute(
       parentNavigatorKey: rootNavKey,
-      name: Routes.chatroom.name,
-      path: Routes.chatroom.route,
-      redirect: authGuardRedirect,
-      pageBuilder: (context, state) {
-        return NoTransitionPage(
-          key: state.pageKey,
-          child: const RoomPage(),
-        );
-      },
-      routes: <RouteBase>[
-        GoRoute(
-          parentNavigatorKey: rootNavKey,
-          path: 'profile',
-          name: Routes.chatProfile.name,
-          redirect: authGuardRedirect,
-          builder: (context, state) => const RoomProfilePage(),
-        ),
-      ],
-    ),
-
-    GoRoute(
-      parentNavigatorKey: rootNavKey,
       name: Routes.createChat.name,
       path: Routes.createChat.route,
       pageBuilder: (context, state) {
@@ -526,6 +504,7 @@ List<RouteBase> makeRoutes(Ref ref) {
         ),
 
         GoRoute(
+          parentNavigatorKey: shellNavKey,
           name: Routes.chat.name,
           path: Routes.chat.route,
           redirect: authGuardRedirect,
@@ -535,6 +514,25 @@ List<RouteBase> makeRoutes(Ref ref) {
               child: const ChatPage(),
             );
           },
+        ),
+
+        GoRoute(
+          name: Routes.chatroom.name,
+          path: Routes.chatroom.route,
+          redirect: authGuardRedirect,
+          pageBuilder: (context, state) {
+            return NoTransitionPage(
+              key: state.pageKey,
+              child: const RoomPage(),
+            );
+          },
+        ),
+
+        GoRoute(
+          name: Routes.chatProfile.name,
+          path: Routes.chatProfile.route,
+          redirect: authGuardRedirect,
+          builder: (context, state) => const RoomProfilePage(),
         ),
 
         GoRoute(
@@ -550,15 +548,14 @@ List<RouteBase> makeRoutes(Ref ref) {
         ),
 
         // ---- SETTINGS
-
         GoRoute(
-          name: Routes.info.name,
-          path: Routes.info.route,
+          name: Routes.settings.name,
+          path: Routes.settings.route,
           redirect: authGuardRedirect,
           pageBuilder: (context, state) {
             return NoTransitionPage(
               key: state.pageKey,
-              child: const SettingsInfoPage(),
+              child: const SettingsMenuPage(),
             );
           },
         ),
@@ -573,19 +570,6 @@ List<RouteBase> makeRoutes(Ref ref) {
             );
           },
         ),
-
-        GoRoute(
-          name: Routes.settings.name,
-          path: Routes.settings.route,
-          redirect: authGuardRedirect,
-          pageBuilder: (context, state) {
-            return NoTransitionPage(
-              key: state.pageKey,
-              child: const SettingsMenuPage(),
-            );
-          },
-        ),
-
         GoRoute(
           name: Routes.settingsLabs.name,
           path: Routes.settingsLabs.route,
@@ -594,6 +578,17 @@ List<RouteBase> makeRoutes(Ref ref) {
             return NoTransitionPage(
               key: state.pageKey,
               child: const SettingsLabsPage(),
+            );
+          },
+        ),
+        GoRoute(
+          name: Routes.info.name,
+          path: Routes.info.route,
+          redirect: authGuardRedirect,
+          pageBuilder: (context, state) {
+            return NoTransitionPage(
+              key: state.pageKey,
+              child: const SettingsInfoPage(),
             );
           },
         ),
@@ -731,7 +726,7 @@ List<RouteBase> makeRoutes(Ref ref) {
             if (authGuarded != null) {
               return authGuarded;
             }
-            if (context.mounted && isDesktop(context)) {
+            if (context.mounted && isDesktop) {
               return Routes.dashboard.route;
             } else {
               return Routes.updates.route;
